Move parquet specific tests into dedicated file
